refactor(settings): make delete-all atomic (Greptile #745 round 2)
The delete and the balance settlement around it ran as three separate database
transactions, so a crash between the DELETE and the recompute could leave the
transactions gone but a manual account still displaying a total for them.

Extracts DeleteAllTransactionsUseCase, which wraps ensure-opening → delete →
recompute in a single withTransaction, mirroring DeleteTransactionUseCase. The
ViewModel keeps only the UI concerns: progress state, widget refresh, result
message.
